Complete 'EventsQueue' class
